Move autoimpl impls to sub-modules
